LRQA-40401 Rename all classes from "TestClassGroup" to "TestGroup"
