+++ modules/test/jenkins-results-parser/src/main/java/com/liferay/jenkins/results/parser/BaseTestGroup.java
@@ -22,7 +22,7 @@ import java.util.List;
 /**
  * @author Peter Yoo
  */
-public abstract class BaseTestClassGroup implements TestClassGroup {
+public abstract class BaseTestGroup implements TestGroup {
 
 	public List<File> getTestClassFiles() {
 		return testClassFiles;
