+++ modules/test/jenkins-results-parser/src/main/java/com/liferay/jenkins/results/parser/PluginsBatchTestGroup.java
@@ -29,48 +29,56 @@ import java.nio.file.attribute.BasicFileAttributes;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
+import java.util.Properties;
 
 /**
  * @author Michael Hashimoto
  */
-public class TCKJunitBatchTestClassGroup extends BatchTestClassGroup {
+public class PluginsBatchTestGroup extends BatchTestGroup {
 
-	protected TCKJunitBatchTestClassGroup(
+	protected PluginsBatchTestGroup(
 		String batchName, PortalGitWorkingDirectory portalGitWorkingDirectory,
 		String testSuiteName) {
 
 		super(batchName, portalGitWorkingDirectory, testSuiteName);
 
-		File workingDirectory = portalGitWorkingDirectory.getWorkingDirectory();
+		Properties portalReleaseProperties =
+			JenkinsResultsParserUtil.getProperties(
+				new File(
+					portalGitWorkingDirectory.getWorkingDirectory(),
+					"release.properties"));
 
-		File tckHomeDirectory = new File(workingDirectory, "tools/tck");
-
-		if (!tckHomeDirectory.exists()) {
-			tckHomeDirectory = new File(
+		try {
+			_pluginsGitWorkingDirectory = new PluginsGitWorkingDirectory(
+				portalGitWorkingDirectory.getUpstreamBranchName(),
 				JenkinsResultsParserUtil.getProperty(
-					portalTestProperties, "tck.home"));
+					portalReleaseProperties, "lp.plugins.dir"));
+
+			_pluginNamesExcludePathMatchers = _getPluginNamesPathMatchers(
+				"test.batch.plugin.names.excludes");
+			_pluginNamesIncludePathMatchers = _getPluginNamesPathMatchers(
+				"test.batch.plugin.names.includes");
+
+			setTestClassFiles();
+
+			setAxisTestGroups();
 		}
-
-		_tckHomeDirectory = tckHomeDirectory;
-
-		_testClassNameExcludePathMatchers = _getTestClassNamesPathMatchers(
-			"test.batch.class.names.excludes");
-		_testClassNameIncludePathMatchers = _getTestClassNamesPathMatchers(
-			"test.batch.class.names.includes");
-
-		setTestClassFiles();
-
-		setAxisTestClassGroups();
+		catch (IOException ioe) {
+			throw new RuntimeException(ioe);
+		}
 	}
 
 	protected void setTestClassFiles() {
+		File workingDirectory =
+			_pluginsGitWorkingDirectory.getWorkingDirectory();
+
 		try {
 			Files.walkFileTree(
-				_tckHomeDirectory.toPath(),
+				workingDirectory.toPath(),
 				new SimpleFileVisitor<Path>() {
 
 					@Override
-					public FileVisitResult visitFile(
+					public FileVisitResult preVisitDirectory(
 							Path filePath, BasicFileAttributes attrs)
 						throws IOException {
 
@@ -78,8 +86,20 @@ public class TCKJunitBatchTestClassGroup extends BatchTestClassGroup {
 							return FileVisitResult.SKIP_SUBTREE;
 						}
 
-						if (_pathIncluded(filePath)) {
-							testClassFiles.add(filePath.toFile());
+						return FileVisitResult.CONTINUE;
+					}
+
+					@Override
+					public FileVisitResult visitFile(
+							Path filePath, BasicFileAttributes attrs)
+						throws IOException {
+
+						if (_pathIncluded(filePath) &&
+							!_pathExcluded(filePath)) {
+
+							File file = filePath.toFile();
+
+							testClassFiles.add(file.getParentFile());
 						}
 
 						return FileVisitResult.CONTINUE;
@@ -87,12 +107,12 @@ public class TCKJunitBatchTestClassGroup extends BatchTestClassGroup {
 
 					private boolean _pathExcluded(Path path) {
 						return _pathMatches(
-							path, _testClassNameExcludePathMatchers);
+							path, _pluginNamesExcludePathMatchers);
 					}
 
 					private boolean _pathIncluded(Path path) {
 						return _pathMatches(
-							path, _testClassNameIncludePathMatchers);
+							path, _pluginNamesIncludePathMatchers);
 					}
 
 					private boolean _pathMatches(
@@ -112,46 +132,46 @@ public class TCKJunitBatchTestClassGroup extends BatchTestClassGroup {
 		catch (IOException ioe) {
 			throw new RuntimeException(
 				"Unable to search for test file names in " +
-					_tckHomeDirectory.getPath(),
+					workingDirectory.getPath(),
 				ioe);
 		}
 
 		Collections.sort(testClassFiles);
 	}
 
-	private List<PathMatcher> _getTestClassNamesPathMatchers(
-		String propertyName) {
+	private List<PathMatcher> _getPluginNamesPathMatchers(String propertyName) {
+		String pluginNamesRelativeGlobs = getFirstPropertyValue(propertyName);
 
-		String testClassNamesRelativeGlobs = getFirstPropertyValue(
-			propertyName);
+		if ((pluginNamesRelativeGlobs == null) ||
+			pluginNamesRelativeGlobs.isEmpty()) {
 
-		if ((testClassNamesRelativeGlobs == null) ||
-			testClassNamesRelativeGlobs.isEmpty()) {
-
-			return Collections.emptyList();
+			return new ArrayList<>();
 		}
 
 		List<PathMatcher> pathMatchers = new ArrayList<>();
 
-		String workingDirectoryPath = _tckHomeDirectory.getAbsolutePath();
+		File workingDirectory =
+			_pluginsGitWorkingDirectory.getWorkingDirectory();
 
-		FileSystem fileSystem = FileSystems.getDefault();
+		String workingDirectoryPath = workingDirectory.getAbsolutePath();
 
-		for (String testClassNamesRelativeGlob :
-				testClassNamesRelativeGlobs.split(",")) {
+		for (String pluginNamesRelativeGlob :
+				pluginNamesRelativeGlobs.split(",")) {
+
+			FileSystem fileSystem = FileSystems.getDefault();
 
 			pathMatchers.add(
 				fileSystem.getPathMatcher(
 					JenkinsResultsParserUtil.combine(
 						"glob:", workingDirectoryPath, "/",
-						testClassNamesRelativeGlob)));
+						pluginNamesRelativeGlob)));
 		}
 
 		return pathMatchers;
 	}
 
-	private final File _tckHomeDirectory;
-	private final List<PathMatcher> _testClassNameExcludePathMatchers;
-	private final List<PathMatcher> _testClassNameIncludePathMatchers;
+	private final List<PathMatcher> _pluginNamesExcludePathMatchers;
+	private final List<PathMatcher> _pluginNamesIncludePathMatchers;
+	private final PluginsGitWorkingDirectory _pluginsGitWorkingDirectory;
 
 }
