+++ modules/test/jenkins-results-parser/src/main/java/com/liferay/jenkins/results/parser/TestGroup.java
@@ -16,35 +16,13 @@ package com.liferay.jenkins.results.parser;
 
 import java.io.File;
 
+import java.util.List;
+
 /**
  * @author Michael Hashimoto
  */
-public class AxisTestClassGroup extends BaseTestClassGroup {
+public interface TestGroup {
 
-	public String getBatchName() {
-		return _batchTestClassGroup.getBatchName();
-	}
-
-	public BatchTestClassGroup getBatchTestClassGroup() {
-		return _batchTestClassGroup;
-	}
-
-	public int getId() {
-		return _id;
-	}
-
-	protected AxisTestClassGroup(
-		BatchTestClassGroup batchTestClassGroup, int id) {
-
-		_batchTestClassGroup = batchTestClassGroup;
-		_id = id;
-	}
-
-	protected void addTestClassFile(File testClassFile) {
-		testClassFiles.add(testClassFile);
-	}
-
-	private final BatchTestClassGroup _batchTestClassGroup;
-	private final int _id;
+	public List<File> getTestClassFiles();
 
 }
